day_18/requirements.txt
colorgram.py==1.2.0
pillow==10.4.0
